agents/workflow_agent/requirements.txt
python-dotenv
asyncio
langchain-google-genai
google-generativeai
pydantic
browser-use
